brain.py
import json
import os
from groq import Groq
from dotenv import load_dotenv

load_dotenv()

_client = Groq(api_key=os.getenv("GROQ_API_KEY"))

_SYSTEM_PROMPT = """Você é o cérebro do JARVIS, assistente pessoal do senhor Schimidt.
Analise o comando de voz transcrito e retorne APENAS um JSON válido com esta estrutura:
{"intent": "string", "entities": {}, "response": "string"}

Intents disponíveis:
- "weather"         → clima/tempo
- "search"          → pesquisar (entities: {"query": "..."})
- "youtube"         → abrir YouTube
- "github"          → abrir GitHub
- "whatsapp"        → abrir WhatsApp
- "work_mode"       → modo trabalho (landing page + VS Code)
- "time"            → hora atual
- "date"            → data atual
- "system_stats"    → status CPU/RAM/bateria
- "lock"            → trancar o computador
- "restart"         → reiniciar o sistema
- "mute"            → silenciar áudio
- "volume_up"       → aumentar volume
- "volume_down"     → diminuir volume
- "minimize"        → minimizar janelas
- "screenshot"      → tirar print da tela
- "open_downloads"  → pasta downloads
- "open_documents"  → pasta documentos
- "open_desktop"    → área de trabalho
- "open_notepad"    → bloco de notas
- "open_calc"       → calculadora
- "open_vscode"     → Visual Studio Code
- "open_taskmgr"    → gerenciador de tarefas
- "mode_focus"      → modo foco (fecha distrações, abre VS Code)
- "mode_relax"      → modo relaxar (abre YouTube)
- "clear_console"   → limpar console
- "project_padaria" → Padaria Santo Antônio
- "project_sukhafe" → Sukhafé
- "greeting"        → saudação ou resposta ao ser chamado
- "unknown"         → comando não reconhecido

"response" deve ser em português, estilo JARVIS: formal, profissional, levemente sarcástico.
Use variações — nunca repita a mesma frase duas vezes seguidas.
Responda APENAS com o JSON. Nenhum markdown, nenhum texto fora do JSON."""


def understand(text: str) -> dict:
    try:
        completion = _client.chat.completions.create(
            model="llama-3.1-8b-instant",
            messages=[
                {"role": "system", "content": _SYSTEM_PROMPT},
                {"role": "user", "content": text},
            ],
            temperature=0.7,
            max_tokens=256,
            response_format={"type": "json_object"},
        )
        return json.loads(completion.choices[0].message.content)
    except Exception as e:
        print(f"[BRAIN] Erro: {e}")
        return {
            "intent": "unknown",
            "entities": {},
            "response": "Tive uma falha de processamento, senhor. Pode repetir?",
        }
