requirements.txt
groq==1.2.0
colorama==0.4.6
comtypes==1.4.16
Flask==3.1.3
Flask-SocketIO==5.6.1
geocoder==1.38.1
numpy==2.4.4
psutil==7.2.2
PyAutoGUI==0.9.54
pycaw==20251023
pydub==0.25.1
python-dotenv==1.2.2
requests==2.33.1
sounddevice==0.5.5
vosk==0.3.45
